Moved RatioVertex to subpackage since it is used by both the Greedy vertex cover alg and the ClarksonTwoApprox
