+++ jgrapht-core/src/main/java/org/jgrapht/alg/vertexcover/BarYehudaEvenTwoApproxVCImpl.java
@@ -49,10 +49,13 @@ import java.util.Set;
  * Implementation of the 2-opt algorithm for a minimum weighted vertex cover by
  * R. Bar-Yehuda and S. Even. A linear time approximation algorithm for the weighted vertex cover problem. J. of
  * Algorithms 2:198-203, 1981.
- * The solution is guaranteed to be within 2 times the optimum solution. Runtime: O(|E|)
+ * The solution is guaranteed to be within 2 times the optimum solution.
  * An easier-to-read version of this algorithm can be found here: <a href="https://www.cs.umd.edu/class/spring2011/cmsc651/vc.pdf">https://www.cs.umd.edu/class/spring2011/cmsc651/vc.pdf/a>
  *
  * Note: this class supports pseudo-graphs
+ * Runtime: O(|E|)
+ *
+ * TODO: Remove the UndirectedSubgraph dependency! Querying vertex degrees on these graphs is actually slow! This does affect the runtime complexity. Better would be to just work on a clone of the original graph!
  *
  * @author Joris Kinable
  *
@@ -65,7 +68,7 @@ public class BarYehudaEvenTwoApproxVCImpl<V,E> implements MinimumWeightedVertexC
         Set<V> cover=new LinkedHashSet<>();
         double weight=0;
         UndirectedGraph<V,E> copy= new UndirectedSubgraph<>(graph, null, null);
-//        UndirectedGraph<V,E> copy=graph.clone();
+//        UndirectedGraph<V,E> copy=(UndirectedGraph<V,E>)graph.clone();
         Map<V, Double> W=new HashMap<>();
         for(V v : graph.vertexSet())
             W.put(v, vertexWeightMap.get(v));
