+++ jgrapht-core/src/main/java/org/jgrapht/alg/vertexcover/GreedyVCImpl.java
@@ -25,24 +25,20 @@
  * (C) Copyright 2003-2008, by Linda Buisman and Contributors.
  *
  * Original Author:  Joris Kinable
- * Contributor(s):   Barak Naveh
- *                   Christian Hammer
- *                   Linda Buisman
+ * Contributor(s):   -
  *
  * $Id$
  *
  * Changes
  * -------
- * 06-Nov-2003 : Initial revision (LB);
- * 07-Jun-2005 : Made generic (CH);
- * 28-Jul-2016 : Moved to dedicated package; Added greedy implementation for Weighted VC (JK)
- * 31-Jul-2016 : Replaced original implementation (O(|E|*|N|)) by Linda Buisman with more efficient (O(|E|*log|N|)) implementation
+ * 31-Jul-2016 : Initial revision (JK);
  *
  */
 package org.jgrapht.alg.vertexcover;
 
 import org.jgrapht.UndirectedGraph;
 import org.jgrapht.alg.interfaces.MinimumWeightedVertexCoverAlgorithm;
+import org.jgrapht.alg.vertexcover.util.RatioVertex;
 
 import java.util.*;
 
@@ -60,7 +56,7 @@ import java.util.*;
  *
  *
  *
- * @author Linda Buisman
+ * @author Joris Kinable
  * @since Nov 6, 2003
  */
 public class GreedyVCImpl<V,E> implements MinimumWeightedVertexCoverAlgorithm<V,E> {
@@ -82,32 +78,32 @@ public class GreedyVCImpl<V,E> implements MinimumWeightedVertexCoverAlgorithm<V,
         Set<V> cover=new LinkedHashSet<>();
         double weight=0;
 
-        //Create working graph: for every vertex, create a ComparableVertex which maintains its own list of neighbors
-        Map<V, ComparableVertex> vertexEncapsulationMap=new HashMap<>();
-        graph.vertexSet().stream().filter(v -> graph.degreeOf(v) > 0).forEach(v -> vertexEncapsulationMap.put(v, new ComparableVertex(v, vertexWeightMap.get(v))));
+        //Create working graph: for every vertex, create a RatioVertex which maintains its own list of neighbors
+        Map<V, RatioVertex<V>> vertexEncapsulationMap=new HashMap<>();
+        graph.vertexSet().stream().filter(v -> graph.degreeOf(v) > 0).forEach(v -> vertexEncapsulationMap.put(v, new RatioVertex<>(vertexCounter++, v, vertexWeightMap.get(v))));
 
         for(E e : graph.edgeSet()){
             V u=graph.getEdgeSource(e);
-            ComparableVertex ux=vertexEncapsulationMap.get(u);
+            RatioVertex<V> ux=vertexEncapsulationMap.get(u);
             V v=graph.getEdgeTarget(e);
-            ComparableVertex vx=vertexEncapsulationMap.get(v);
+            RatioVertex<V> vx=vertexEncapsulationMap.get(v);
             ux.addNeighbor(vx);
             vx.addNeighbor(ux);
 
             assert(ux.neighbors.get(vx) == vx.neighbors.get(ux)): " in an undirected graph, if vx is a neighbor of ux, then ux must be a neighbor of vx";
         }
 
-        TreeSet<ComparableVertex> workingGraph=new TreeSet<>();
+        TreeSet<RatioVertex<V>> workingGraph=new TreeSet<>();
         workingGraph.addAll(vertexEncapsulationMap.values());
         assert(workingGraph.size() == graph.vertexSet().size());
 
         while(!workingGraph.isEmpty()) { //Continue until all edges are covered
 
             //Find a vertex vx for which W(vx)/degree(vx) is minimal
-            ComparableVertex vx = workingGraph.pollFirst();
+            RatioVertex<V> vx = workingGraph.pollFirst();
             assert(workingGraph.parallelStream().allMatch(ux -> vx.getRatio() <= ux.getRatio())) : "vx does not have the smallest ratio among all elements. VX: "+vx+" WorkingGraph: "+workingGraph;
 
-            for(ComparableVertex nx : vx.neighbors.keySet()){
+            for(RatioVertex<V> nx : vx.neighbors.keySet()){
 
                 if(nx ==vx) //Ignore self loops
                     continue;
@@ -117,7 +113,7 @@ public class GreedyVCImpl<V,E> implements MinimumWeightedVertexCoverAlgorithm<V,
                 //Delete vx from nx' neighbor list. Delete nx from the graph and place it back, thereby updating the ordering of the graph
                 nx.removeNeighbor(vx);
 
-                if (nx.degree > 0)
+                if (nx.getDegree() > 0)
                     workingGraph.add(nx);
 
             }
@@ -131,77 +127,5 @@ public class GreedyVCImpl<V,E> implements MinimumWeightedVertexCoverAlgorithm<V,
         return new VertexCover<>(cover, weight);
     }
 
-    public class ComparableVertex implements Comparable<ComparableVertex>{
-        /** original vertex **/
-        public final V v;
 
-        /** weight of the vertex **/
-        public double weight;
-
-        /** unique id, used to guarantee that compareTo never returns 0 **/
-        public final int ID;
-
-        /** degree of this vertex **/
-        private int degree=0;
-
-        /** Map of neighbors, and a count of the number of edges to this neighbor **/
-        public Map<ComparableVertex, Integer> neighbors;
-
-        public ComparableVertex(V v, double weight){
-            this.ID=vertexCounter++;
-            this.v=v;
-            this.weight=weight;
-            neighbors=new LinkedHashMap<>();
-        }
-
-        public void addNeighbor(ComparableVertex v){
-            if(!neighbors.containsKey(v))
-                neighbors.put(v,1);
-            else
-                neighbors.put(v, neighbors.get(v) + 1);
-            degree++;
-
-            assert(neighbors.values().stream().mapToInt(Integer::intValue).sum() == degree);
-        }
-
-        public void removeNeighbor(ComparableVertex v){
-            degree-=neighbors.get(v);
-            neighbors.remove(v);
-        }
-
-        public double getRatio(){
-            return weight/degree;
-        }
-
-        @Override
-        public int compareTo(ComparableVertex other) {
-            if(this.ID == other.ID) //Same vertex
-                return 0;
-            int result=Double.compare(this.getRatio(), other.getRatio());
-            if(result == 0 ) //If vertices have the same value, resolve tie by an ID comparison
-                return Integer.compare(this.ID, other.ID);
-            else
-                return result;
-        }
-
-        @Override
-        public int hashCode(){
-            return ID;
-        }
-
-        @Override
-        public boolean equals(Object o){
-            if(this==o)
-                return true;
-            else if(!(o instanceof ClarksonTwoApproxVCImpl.ComparableVertex))
-                return false;
-            ComparableVertex other=(ComparableVertex)o;
-            return this.ID==other.ID;
-        }
-
-        @Override
-        public String toString(){
-            return "v"+ID+"("+degree+")";
-        }
-    }
 }
