+++ jgrapht-core/src/test/java/org/jgrapht/alg/util/VertexDegreeComparatorTest.java
@@ -1,3 +1,39 @@
+/* ==========================================
+ * JGraphT : a free Java graph-theory library
+ * ==========================================
+ *
+ * Project Info:  http://jgrapht.sourceforge.net/
+ * Project Creator:  Barak Naveh (http://sourceforge.net/users/barak_naveh)
+ *
+ * (C) Copyright 2003-2008, by Barak Naveh and Contributors.
+ *
+ * This program and the accompanying materials are dual-licensed under
+ * either
+ *
+ * (a) the terms of the GNU Lesser General Public License version 2.1
+ * as published by the Free Software Foundation, or (at your option) any
+ * later version.
+ *
+ * or (per the licensee's choosing)
+ *
+ * (b) the terms of the Eclipse Public License v1.0 as published by
+ * the Eclipse Foundation.
+ */
+/* -----------------
+ * VertexDegreeComparatorTest.java
+ * -----------------
+ * (C) Copyright 2016, by Joris Kinable and Contributors.
+ *
+ * Original Author:  Joris Kinable
+ * Contributor(s):   -
+ *
+ * $Id$
+ *
+ * Changes
+ * -------
+ * 28-Jul-2016 : Initial revision (JK);
+ *
+ */
 package org.jgrapht.alg.util;
 
 import junit.framework.TestCase;
@@ -13,7 +49,9 @@ import java.util.HashMap;
 import java.util.List;
 
 /**
- * Created by jkinable on 7/31/16.
+ * Unit tests for VertexDegreeComparator
+ *
+ * @author Joris Kinable
  */
 public class VertexDegreeComparatorTest extends TestCase {
